requirements.txt
astroid==2.11.2
attrs==21.4.0
awsebcli==3.20.3
bcrypt==3.2.0
blessed==1.19.1
boto3==1.21.23
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.3
cryptography==36.0.2
dill==0.3.4
distlib==0.3.4
Django==2.1.15
django-crispy-forms==1.14.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
filelock==3.6.0
Flask==2.0.3
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
future==0.16.0
idna==3.3
importlib-metadata==4.11.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.7.0
numpy==1.22.3
pandas==1.4.1
paramiko==2.10.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
pycparser==2.21
PyLink==0.3.3
pylint==2.13.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyMySQL==1.0.2
PyNaCl==1.5.0
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.2
semantic-version==2.8.5
six==1.14.0
SQLAlchemy==1.4.31
termcolor==1.1.0
texttable==1.6.4
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.1.1
urllib3==1.26.9
virtualenv==20.13.1
wcwidth==0.1.9
websocket-client==0.59.0
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.8.0
